Refactoring
Move common functions into pkg folder. These functions can be reused for future integration key-expiration-notifier
